Move Pre Istio 1.5 check
