+++ istio/datadog_checks/istio/constants.py
@@ -1,0 +1,12 @@
+# (C) Datadog, Inc. 2020 - Present
+# All rights reserved
+# Licensed under Simplified BSD License (see LICENSE)
+
+## Istio namespaces
+MIXER_NAMESPACE = 'istio.mixer'
+MESH_NAMESPACE = 'istio.mesh'
+PILOT_NAMESPACE = 'istio.pilot'
+GALLEY_NAMESPACE = 'istio.galley'
+CITADEL_NAMESPACE = 'istio.citadel'
+
+DEFAULT_METRIC_LIMIT = 0
